backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
google-auth>=2.22.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.21
asyncpg>=0.28.0
psycopg2-binary>=2.9.7
python-jose>=3.3.0
passlib>=1.7.4
httpx>=0.24.1
celery>=5.3.4
redis>=5.0.1
boto3>=1.28.55
sendgrid>=6.10.0
openai>=1.2.0
python-multipart>=0.0.6
mermaid-cli>=0.1.5
elevenlabs>=0.2.24
aiofiles>=23.2.1